+++ old_src/exploration_env.py
@@ -1,28 +1,27 @@
 from collections import defaultdict
 from copy import deepcopy
 import sys
+from turtle import goto
 # -- local import
 from env_graphs import WeightedGraphEnv
 from env_clock import ExplorationClock
 """
-This file handles the entire environment and keeps a track of the explored
+This file handles the entire environment and keeps track of the explored
 graph. It provides the agent with the current state of the robot.
-It keeps a track of the visited and seen nodes, through which it can create
+It keeps track of the visited and seen nodes, through which it can create
 this state".
 
 So to write it in steps:
- 1. Read the file which contains the edges, and nodes, and make the ground 
-    truth and pruned graph
- 2. Initialize the explored graph
- 2. Initialize the robot to a position
+ 1. Read the file which contains the edges, and nodes, and makes the ground 
+    truth graph.
+ 2. Initialize the explored graph.
+ 2. Initialize the robot to a position.
     Consider the neighboring vertices as seen, and the current vertex as visited.
- 3. Create the state of robot considering local and remote vertices in each direction. 
-    Return to the agent a set of local and remote vertices, each towards a neighboring vertex. So if a vertex has 5 neighbors, it should report this information:
-     a. Local: (Neighboring Node, Cost)
-        Remote: (Neighboring Node, Remote Node, Cost)
-        To represent this information we write:
-        Neighboring_Node: {'lcost':local_cost, 'r_node':Remote_Node, 'r_cost:',remote_cost}
-    b. the agent returns the neighboring node which the agent plans to head towards.
+ 3. Create the state of robot:
+    a. Current robot position
+    b. discovered states: dictionary key: Neighboring Node, val: path cost
+    c. exploration_clock: 
+    a. Discovered vertices: 
  4. Update the explored graph, 
 """
 
@@ -32,7 +31,7 @@ class GraphExplorationEnvironment:
                  edgeFileName,
                  vertexTagFileName,
                  vertexPriorityFileName,
-                 fixed_tw, fixed_tb,
+                 t_a, t_r0,
                  initial_robot_position):
         
         # -- set up filenames
@@ -50,60 +49,41 @@ class GraphExplorationEnvironment:
         self.env.set_up_initial_robot_position(self.initial_robot_position)
         
         # -- set up timers
-        self.tw = fixed_tw
-        self.tb = fixed_tb
+        self.t_a = t_a # time before alarm
+        self.t_r0 = t_r0 # deadline time
         self._set_up_clock()
         
         # -- set up cost
         self.cost_incurred = 0
     
     def _set_up_clock(self):
-        self.exploration_clock = ExplorationClock(self.tw, self.tb)
+        self.exploration_clock = ExplorationClock(self.t_a, self.t_r0)
     
 
     def get_state_consolidated(self):
         cur_pos = self.current_robot_position
 
         # -- the current state is a combination of local states and remote nodes towards a direction
-        local_states = self.env.get_local_unvisited_nodes_gt(cur_pos)
-        print ("EXPLORATION ENV: GET_STATE_CONSOLIDATED: Local states: ", local_states)
-        local_states = self._convert_local_states_list_to_dict(local_states)
-        print ("EXPLORATION ENV: GET_STATE_CONSOLIDATED: Local state dictionary: ", local_states)
-        remote_states = self.env.get_remote_nodes(cur_pos)
-        print ("EXPLORATION ENV: GET_STATE_CONSOLIDATED: Remote states: ", remote_states)
+        # local_states = self.env.get_local_unvisited_nodes_gt(cur_pos)
+        # print ("EXPLORATION ENV: GET_STATE_CONSOLIDATED: Local states: ", local_states)
+        # local_states = self._convert_local_states_list_to_dict(local_states)
+        # print ("EXPLORATION ENV: GET_STATE_CONSOLIDATED: Local state dictionary: ", local_states)
+        # remote_states = self.env.get_remote_nodes(cur_pos)
+        # print ("EXPLORATION ENV: GET_STATE_CONSOLIDATED: Remote states: ", remote_states)
         
         # -- get all discovered states
         discovered_states = self.env.get_discovered_nodes(cur_pos)
         self.discovered_states, self.source_discovered_states = discovered_states
         print ("EXPLORATION ENV: GET_STATE_CONSOLIDATED: Discovered states: ", discovered_states)
-        # input("Continue?")
 
         # -- save to class variables
-        self.local_states = local_states
-        self.remote_states, self.source_remote_states = remote_states
+        # self.local_states = local_states
+        # self.remote_states, self.source_remote_states = remote_states
         
         # -- return the state
         return (cur_pos, self.discovered_states, self.exploration_clock.get_time_limit(), self.cost_incurred)
         
 
-    # def get_state(self):
-    #     cur_pos = self.current_robot_position
-        
-    #     # -- the current state is a combination of local states and remote nodes towards a direction
-    #     local_states = self.env.get_local_unvisited_nodes_gt(cur_pos)
-    #     print ("Local states: ", local_states)
-    #     local_states = self._convert_local_states_list_to_dict(local_states)
-    #     print ("Local state dictionary: ", local_states)
-    #     remote_states = self.env.get_remote_nodes(cur_pos)
-    #     print ("Remote states: ", remote_states)
-        
-    #     # -- save to class variables
-    #     self.local_states = local_states
-    #     self.remote_states, self.source_remote_states = remote_states
-        
-    #     # -- return the state
-    #     return (cur_pos, local_states, self.remote_states, self.exploration_clock.get_time_limit(), self.cost_incurred)
-    
     def _convert_local_states_list_to_dict(self, local_states):
         local_state_dict = {}
         for node in local_states:
@@ -132,18 +112,22 @@ class GraphExplorationEnvironment:
             print("EXPLORATION ENV: step :: Robot is trying to back to the same node ", goto_node)
             time_remaining = self.exploration_clock.update_time_limit()
             return self.get_state_consolidated()
-        elif goto_node in self.local_states.keys():
-            # -- robot is trying to reach a local node
-            print("EXPLORATION ENV: step :: Robot headed for local node")
-            node_reached, cost = self.env.go_local_node(goto_node)
-            # -- increment total cost
-            self.cost_incurred += cost
-        elif goto_node in self.remote_states.keys():
-            # -- robot is trying to reach a remote node
-            print("EXPLORATION ENV: step :: Robot headed for remote node {}".format(goto_node))
-            node_reached, cost = self.env.go_remote_node(self.current_robot_position,goto_node)
-            # -- increment total cost
-            self.cost_incurred += cost
+        elif goto_node in self.discovered_states.keys():
+            # -- robot is trying to reach a discovered node
+            print(f"EXPLORATION ENV: step :: Robot headed for discovered node: {goto_node}")
+            if self.current_robot_position == self.source_discovered_states[goto_node]:
+                # -- robot is trying to reach a local node
+                print("EXPLORATION ENV: step :: Robot headed for local node")
+                node_reached, cost = self.env.go_local_node(goto_node)
+                # -- increment total cost
+                self.cost_incurred += cost
+            else:
+                # -- robot is trying to reach a remote node
+                print("EXPLORATION ENV: step :: Robot headed for remote node {}".format(goto_node))
+                input("EXPLORATION_ENV::STEP::Continue?")
+                node_reached, cost = self.env.go_remote_node(self.current_robot_position,goto_node)
+                # -- increment total cost
+                self.cost_incurred += cost
         elif goto_node == self.initial_robot_position:
             # -- robot is trying to reach home
             print("EXPLORATION ENV: step :: Robot headed for home node")
@@ -200,9 +184,3 @@ class GraphExplorationEnvironment:
             print ("EXPLORATION ENV: get_unvisited_distances_from_node {}".format(node))
             print ("\t Distance to unvisited nodes: {}".format(dist_unvisited_nodes))
         return deepcopy(dist_unvisited_nodes)
-    
-    
-        
-        
-        
-