+++ old_src/changes.md
@@ -5,8 +5,7 @@ Files:
 4. priority_exploration_agent.py
 
 # TODO
-1. Remove remote and local notes and just keep one. At least for the interface that is available to others. (done partially)
-2. What's the best way to consider the remote nodes from Up top? Write wrapper codes till basic ones are changed (2 hours)
+1. Remove remote and local notes and just keep one. At least for the interface that is available to others. (done)
 3. Make a function in exploration_env to create a adjacency matrix for the explored graph. 
 4. Write down clear documentation for:
     1. How do get the adjacency list as a dictionary for the explored graph?